Move my code to the existing `repos_merging.go` and `repos_merging_test.go` files.
